fix: adjust export csv GrainsVol and ReceiptMap, logical helpers in customtable
